Feature/add or edit

+++ WeatherForecast/models/src/main/java/com/mauscoelho/data/Main.java
@@ -5,9 +5,9 @@ public class Main {
     public int humidity;
     public float pressure;
     public float temp_max;
-    public int sea_level;
+    public float sea_level;
     public float temp_min;
     public float temp;
     public float grnd_level;
-    public int temp_kf;
+    public float temp_kf;
 }
